src/redshift_agents/requirements.txt
# Core dependencies for Redshift Modernization Agents
# Python 3.12 required

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# AWS CDK (infrastructure-as-code)
aws-cdk-lib>=2.150.0
constructs>=10.0.0

# Logging and monitoring
python-json-logger>=2.0.7

# Environment management
python-dotenv>=1.0.0
